fix(1179): Fix namespace parsing in template tag create
