+++ lib/templateFactory.js
@@ -3,6 +3,8 @@
 const BaseFactory = require('./baseFactory');
 const Template = require('./template');
 const compareVersions = require('compare-versions');
+const hoek = require('hoek');
+const parseTemplateConfigName = require('./helper').parseTemplateConfigName;
 const schema = require('screwdriver-data-schema');
 const TEMPLATE_NAME_REGEX = schema.config.regex.FULL_TEMPLATE_NAME;
 const TEMPLATE_NAME_REGEX_WITH_NAMESPACE = schema.config.regex.FULL_TEMPLATE_NAME_WITH_NAMESPACE;
@@ -125,41 +127,25 @@ class TemplateFactory extends BaseFactory {
     create(config) {
         const [, major, minor] = VERSION_REGEX.exec(config.version);
         const searchVersion = minor ? `${major}${minor}` : major;
-
-        // Set namespace if it doesn't already exist
-        if (!config.namespace) {
-            const slashIndex = config.name.indexOf('/');
-
-            // Use string in front of slash for namespace if there is a slash
-            if (slashIndex > -1) {
-                const [, namespace, name]
-                    = TEMPLATE_NAME_REGEX_WITH_NAMESPACE.exec(config.name);
-
-                config.namespace = namespace;
-                config.name = name;
-            // Set namespace to default if there is no slash
-            } else {
-                config.namespace = 'default';
-            }
-        }
-
-        const fullTemplateName = `${config.namespace}/${config.name}@${searchVersion}`;
+        const nameObj = parseTemplateConfigName(config);
+        const result = hoek.applyToDefaults(config, nameObj);
+        const fullTemplateName = `${result.namespace}/${result.name}@${searchVersion}`;
 
         return this.getTemplate(fullTemplateName)
             .then((latest) => {
                 if (!latest) {
-                    config.version = minor ? `${major}${minor}.0` : `${major}.0.0`;
+                    result.version = minor ? `${major}${minor}.0` : `${major}.0.0`;
                 } else {
                     // eslint-disable-next-line max-len
                     const [, latestMajor, latestMinor, latestPatch] = VERSION_REGEX.exec(latest.version);
                     const patch = parseInt(latestPatch.slice(1), 10) + 1;
 
-                    config.version = `${latestMajor}${latestMinor}.${patch}`;
+                    result.version = `${latestMajor}${latestMinor}.${patch}`;
                 }
 
-                config.createTime = (new Date()).toISOString();
+                result.createTime = (new Date()).toISOString();
 
-                return super.create(config);
+                return super.create(result);
             });
     }
