+++ lib/templateTagFactory.js
@@ -2,6 +2,8 @@
 
 const BaseFactory = require('./baseFactory');
 const TemplateTag = require('./templateTag');
+const hoek = require('hoek');
+const parseTemplateConfigName = require('./helper').parseTemplateConfigName;
 const schema = require('screwdriver-data-schema');
 const TEMPLATE_NAME_REGEX = schema.config.regex.FULL_TEMPLATE_NAME;
 const TEMPLATE_NAME_REGEX_WITH_NAMESPACE = schema.config.regex.FULL_TEMPLATE_NAME_WITH_NAMESPACE;
@@ -105,23 +107,12 @@ class TemplateTagFactory extends BaseFactory {
      * @return {Promise}
      */
     create(config) {
-        config.createTime = (new Date()).toISOString();
+        const nameObj = parseTemplateConfigName(config);
+        const result = hoek.applyToDefaults(config, nameObj);
 
-        if (config.name && !config.namespace) {
-            // eslint-disable-next-line no-underscore-dangle
-            return this._getNameAndNamespace(config.name).then((parsedTemplateName) => {
-                const { namespace, name } = parsedTemplateName;
+        result.createTime = (new Date()).toISOString();
 
-                if (namespace) {
-                    config.namespace = namespace;
-                    config.name = name;
-                }
-
-                return super.create(config);
-            });
-        }
-
-        return super.create(config);
+        return super.create(result);
     }
 
     /**
