+++ package.json
@@ -52,7 +52,7 @@
     "hoek": "^5.0.3",
     "iron": "^5.0.1",
     "lodash": "^4.17.10",
-    "screwdriver-config-parser": "^4.3.1",
+    "screwdriver-config-parser": "^4.4.1",
     "screwdriver-data-schema": "^18.24.7",
     "screwdriver-workflow-parser": "^1.5.0"
   }
